on avance vers une gestion de tous les criteres de l'api.
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@6612 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _test_/FpipR/trunk/base/FpipR_temporaire.php
@@ -23,7 +23,16 @@ $fpipr_field = array(
 	"ispublic"=> "ENUM ('0','1') NOT NULL",
 	"isfriend"=> "ENUM ('0','1') NOT NULL",
 	"isfamily"=> "ENUM ('0','1') NOT NULL",
-	"originalformat" => "char(4) DEFAULT 'jpg'"
+	"originalformat" => "char(4) DEFAULT 'jpg'",
+	"license" => "ENUM ('1','2','3','4','5','6')", //http://flickr.com/services/api/flickr.photos.licenses.getInfo.html
+	"date_upload" => "DATETIME", 
+	"date_taken" => "INT", 
+	"owner_name" => "text DEFAULT '' NOT NULL",
+	"icon_server" => "int NOT NULL",
+	"last_update" => "INT",
+	"latitude" => "FLOAT",
+	"longitude" => "FLOAT",
+	"accuracy" => "smallint NOT NULL"
 );
 
 $fpipr_key = array(
@@ -69,19 +78,22 @@ function FpipR_fill_table($method,$arguments){
 									 $arguments['max_upload_date'], $arguments['min_taken_date'],
 									 $arguments['max_taken_date'], $arguments['license'],
 									 $arguments['sort'], $arguments['privacy_filter'],
-									 $arguments['extras'], $arguments['auth_token']
-									 );
+									 "license,date_upload,date_taken,owner_name,icon_server,original_format,last_update,geo",//on recupere tous les extras possibles
+ $arguments['auth_token']);
+
+
 	 $not_id = '';
 	  foreach($photos->photos as $photo) {
-		$query = "INSERT IGNORE INTO spip_fpipr_photos (id_photo,user_id,secret,server,title,ispublic,isfriend,isfamily,originalformat)";
-		$query .= " VALUES (".intval($photo->id).','.spip_abstract_quote($photo->owner).','.spip_abstract_quote($photo->secret).','.intval($photo->server).','.spip_abstract_quote($photo->title).','.intval($photo->idpublic).','.intval($photo->isfriend).','.intval($photo->isfamily).','.spip_abstract_quote($photo->originalformat).")";
-//		$query .= ' ON DUPLICATE KEY UPDATE counter_usage=counter_usage+1';
+		$query = "INSERT IGNORE INTO spip_fpipr_photos (id_photo,user_id,secret,server,title,ispublic,isfriend,isfamily,originalformat,license,date_upload,date_taken,owner_name,icon_server,original_format,last_update,longitude,latitude,accuracy)";
+		$query .= " VALUES (".intval($photo->id).','.spip_abstract_quote($photo->owner).','.spip_abstract_quote($photo->secret).','.intval($photo->server).','.spip_abstract_quote($photo->title).','.intval($photo->idpublic).','.intval($photo->isfriend).','.intval($photo->isfamily).','.spip_abstract_quote($photo->originalformat).','.intval($photo->license).','.spip_abstract_quote($photo->dateupload).','.intval($photo->datetaken).','.spip_abstract_quote($photo->ownername).','.intval($photo->iconserver).','.intval($photo->lastupdate).','.floatval($photo->longitude).','.floatval($photo->latitude).','.intval($photo->accuracy).")";
 		spip_query($query);
 		$not_id .= ','.intval($photo->id);
 	  }
-	  $not_id = substr($not_id,1);
-	  $query = "DELETE FROM spip_fpipr_photos WHERE id_photo NOT IN ($not_id)";
-	  spip_query($query);
+	  if($not_id) {
+		  $not_id = substr($not_id,1);
+		  $query = "DELETE FROM spip_fpipr_photos WHERE id_photo NOT IN ($not_id)";
+		  spip_query($query);
+	  }
 	  break;
 	default: 
 	  return;
